pyproject.toml
[tool.poetry]
name = "cg-sc23-backend"
version = "0.1.0"
description = ""
authors = ["Sangwon Ko <evbayooooo@gmail.com>"]
packages = [{include = "cg_sc23_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.7"
pydantic = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
